// src/payload/payload-types.ts
export interface Media {
  id: string;
  alt: string;
  caption?: {
    [k: string]: unknown;
  }[];
  updatedAt: string;
  createdAt: string;
  url?: string;
  filename?: string;
  mimeType?: string;
  filesize?: number;
  width?: number;
  height?: number;
}
// ...
export interface Food {
  id: string;
  title: string;
  media?: string | Media;
  updatedAt: string;
  createdAt: string;
}